+++ CHANGELOG.rst
@@ -61,6 +61,8 @@ Fixed
     - Invalid URL in ``build_invite_url()`` caused a crash (#149)
 - **Samplesheets**
     - Delete investigation if import fails (#138)
+- **Timeline**
+    - Fix event id parameter in Taskflow view
 
 Removed
 -------